pyipk: ssh connection handling
+ fixed error handling in pysftp library
+ new remote host dialog in ipkqt
+ url option in ipkcmd

+++ pyipk/libipk/filesource.py
@@ -10,6 +10,9 @@ class FileSource :
 	def write( self , data ) :
 		passe
 
+	def filename( self ) :
+		return ''
+
 	def __del__( self ) :
 		pass
 
@@ -26,24 +29,37 @@ class FS_local( FileSource ) :
 		with open(self.fn,'w') as f :
 			f.write(data)
 
+	def filename( self ) :
+		return self.fn
+
+from miscipk.pysftp import ConnectionError
+from miscipk.pysftp import AuthenticationError
+from miscipk.pysftp import HostNotFoundError
+
 class FS_ssh( FS_local ) :
-	def __init__( self , serv , path , port = 22 ) :
+	def __init__( self , serv , path , *l , **m ) :
+		self.fn = None
 		self.path = path
-		self.con = pysftp.Connection( serv , port = port )
+		self.con = pysftp.Connection( serv , *l , **m )
 		fd , fn  = tempfile.mkstemp('.xml','params-')
 		os.close(fd)
-
 		FS_local.__init__( self , fn )
+		self.refresh()
+
+	def refresh( self ) :
+		self.con.get( self.path , self.fn )
 
 	def read( self ) :
-		self.con.get( self.path , self.fn )
 		return FS_local.read( self )
 
 	def write( self , data ) :
 		FS_local.write( self , data )
 		self.con.put( self.fn , self.path )
 
+	def filename( self ) :
+		return self.fn
+
 	def __del__( self ) :
 		FS_local.__del__( self )
-		os.remove(self.fn)
+		if self.fn != None : os.remove(self.fn)
